cipher-aura-backend/db.py
from pymongo import MongoClient
import os
from dotenv import load_dotenv

load_dotenv()
client = MongoClient(os.getenv("MONGO_URI"))
db = client["cipher_aura"]
users = db["users"]
messages = db["messages"]
